Polish and tested.
There is an issue with the auto run-as.
Super Launcher needs to be able to detect if it was just launched as another user automatically and disable the "auto run-as" settings menu to prevent a user configuring a loop.

Possibly by replacing the menu with text that indicates to the user to re-launch and cancel entering the password to reset the setting.